pyproject.toml
[tool.poetry]
name = "abinbev-case"
version = "0.1.0"
description = "ABInBev Data Engineering Case - Medallion Architecture Pipeline"
authors = ["Michael Santos"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pyspark = "^3.5.0"
delta-spark = "^3.2.0"
pandas = "^2.2.0"
pyarrow = ">=17.0.0"
numpy = ">=1.26.0,<2.0.0"
pyyaml = "^6.0"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"
prometheus-client = "^0.19.0"
requests = "^2.31.0"
setuptools = ">=70.0.0"
keyring = "^25.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
mypy = "^1.5.0"
bandit = "^1.7.0"
jupyter = "^1.1.1"
apache-airflow = "2.10.4"
matplotlib = "^3.10.8"

[tool.poetry.group.observability.dependencies]
opentelemetry-api = "^1.20.0"
opentelemetry-sdk = "^1.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
